Add new projects and refactored code
