+++ src/TFGInfotecAbstractions/Interfaces/IMenuItem.cs
@@ -1,0 +1,11 @@
+﻿namespace TFGInfotecAbstractions.Interfaces
+{
+	public interface IMenuItem
+	{
+		string Name { get; set; }
+		string Description { get; set; }
+		int Rating { get; set; }
+		public double Price { get; set; }
+		public string Image { get; set; }
+	}
+}
