+++ src/TFGInfotecApi/TFGInfotecApi.csproj
@@ -18,4 +18,10 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\TFGInfotecAbstractions\TFGInfotecAbstractions.csproj" />
+    <ProjectReference Include="..\TFGInfotecCore\TFGInfotecCore.csproj" />
+    <ProjectReference Include="..\TFGInfotecInfrastructure\TFGInfotecInfrastructure.csproj" />
+  </ItemGroup>
+
 </Project>
