+++ src/TFGInfotecCore/TFGInfotecCore.csproj
@@ -1,0 +1,14 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\TFGInfotecAbstractions\TFGInfotecAbstractions.csproj" />
+    <ProjectReference Include="..\TFGInfotecInfrastructure\TFGInfotecInfrastructure.csproj" />
+  </ItemGroup>
+
+</Project>
